Separate out module LogitsDecoder
